Can now set/get volume for alsa drivers. Still a problem in config dialog, and with the e_slider tho.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@25727 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e_modules/mixer/alsa_mixer.h
@@ -22,6 +22,6 @@ void        *alsa_card_get_channel  (void *data, int channel_id);
 void         alsa_free_cards        (void *data);
 
 int          alsa_get_volume        (int card_id, int channel_id);
-int          alsa_set_volume        (int card_id, int channel_id, int vol);
+int          alsa_set_volume        (int card_id, int channel_id, double vol);
 
 #endif
